EZP-24329: extract Document Mapper from Handler
