+++ eZ/Publish/Core/Search/Solr/Content/DocumentMapper.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * This file is part of the eZ Publish Kernel package
+ *
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ * @version //autogentag//
+ */
+
+namespace eZ\Publish\Core\Search\Solr\Content;
+
+use eZ\Publish\SPI\Persistence\Content;
+use eZ\Publish\SPI\Persistence\Content\Location;
+use eZ\Publish\SPI\Persistence\Content\Section;
+
+/**
+ * Mapper maps Content and Location objects to a Document object, representing a
+ * document in Solr index storage.
+ *
+ * Note that custom implementations might need to be accompanied by custom mappings.
+ */
+interface DocumentMapper
+{
+    /**
+     * Maps given Content to a Document.
+     *
+     * @param \eZ\Publish\SPI\Persistence\Content $content
+     *
+     * @return \eZ\Publish\SPI\Search\Document[]
+     */
+    public function mapContent( Content $content );
+}
