+++ eZ/Publish/Core/Search/Solr/Content/DocumentMapper/TranslationDocumentMapper.php
@@ -1,0 +1,531 @@
+<?php
+/**
+ * This file is part of the eZ Publish Kernel package
+ *
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ * @version //autogentag//
+ */
+
+namespace eZ\Publish\Core\Search\Solr\Content\DocumentMapper;
+
+use eZ\Publish\Core\Search\Solr\Content\DocumentMapper;
+use eZ\Publish\SPI\Persistence\Content;
+use eZ\Publish\SPI\Persistence\Content\Type as ContentType;
+use eZ\Publish\SPI\Persistence\Content\Location;
+use eZ\Publish\SPI\Persistence\Content\Section;
+use eZ\Publish\SPI\Search\Field;
+use eZ\Publish\SPI\Search\Document;
+use eZ\Publish\SPI\Search\FieldType;
+use eZ\Publish\SPI\Persistence\Content\Location\Handler as LocationHandler;
+use eZ\Publish\SPI\Persistence\Content\Type\Handler as ContentTypeHandler;
+use eZ\Publish\SPI\Persistence\Content\ObjectState\Handler as ObjectStateHandler;
+use eZ\Publish\SPI\Persistence\Content\Section\Handler as SectionHandler;
+use eZ\Publish\Core\Search\Common\FieldRegistry;
+use eZ\Publish\Core\Search\Common\FieldNameGenerator;
+
+/**
+ *
+ */
+class TranslationDocumentMapper implements DocumentMapper
+{
+    /**
+     * Field registry
+     *
+     * @var \eZ\Publish\Core\Search\Common\FieldRegistry
+     */
+    protected $fieldRegistry;
+
+    /**
+     * Location handler
+     *
+     * @var \eZ\Publish\SPI\Persistence\Content\Location\Handler
+     */
+    protected $locationHandler;
+
+    /**
+     * Content type handler
+     *
+     * @var \eZ\Publish\SPI\Persistence\Content\Type\Handler
+     */
+    protected $contentTypeHandler;
+
+    /**
+     * Object state handler
+     *
+     * @var \eZ\Publish\SPI\Persistence\Content\ObjectState\Handler
+     */
+    protected $objectStateHandler;
+
+    /**
+     * Section handler
+     *
+     * @var \eZ\Publish\SPI\Persistence\Content\Section\Handler
+     */
+    protected $sectionHandler;
+
+    /**
+     * Field name generator
+     *
+     * @var \eZ\Publish\Core\Search\Common\FieldNameGenerator
+     */
+    protected $fieldNameGenerator;
+
+    /**
+     * Creates a new content handler.
+     *
+     * @param \eZ\Publish\Core\Search\Common\FieldRegistry $fieldRegistry
+     * @param \eZ\Publish\SPI\Persistence\Content\Location\Handler $locationHandler
+     * @param \eZ\Publish\SPI\Persistence\Content\Type\Handler $contentTypeHandler
+     * @param \eZ\Publish\SPI\Persistence\Content\ObjectState\Handler $objectStateHandler
+     * @param \eZ\Publish\SPI\Persistence\Content\Section\Handler $sectionHandler
+     * @param \eZ\Publish\Core\Search\Common\FieldNameGenerator $fieldNameGenerator
+     */
+    public function __construct(
+        FieldRegistry $fieldRegistry,
+        LocationHandler $locationHandler,
+        ContentTypeHandler $contentTypeHandler,
+        ObjectStateHandler $objectStateHandler,
+        SectionHandler $sectionHandler,
+        FieldNameGenerator $fieldNameGenerator
+    )
+    {
+        $this->fieldRegistry = $fieldRegistry;
+        $this->locationHandler = $locationHandler;
+        $this->contentTypeHandler = $contentTypeHandler;
+        $this->objectStateHandler = $objectStateHandler;
+        $this->sectionHandler = $sectionHandler;
+        $this->fieldNameGenerator = $fieldNameGenerator;
+    }
+
+    /**
+     * Maps given Content to a Document.
+     *
+     * @param \eZ\Publish\SPI\Persistence\Content $content
+     *
+     * @return \eZ\Publish\SPI\Search\Document[]
+     */
+    public function mapContent( Content $content )
+    {
+        $locations = $this->locationHandler->loadLocationsByContent( $content->versionInfo->contentInfo->id );
+        $section = $this->sectionHandler->load( $content->versionInfo->contentInfo->sectionId );
+        $mainLocation = null;
+        $isSomeLocationVisible = false;
+        $locationDocuments = array();
+        foreach ( $locations as $location )
+        {
+            $locationDocuments[] = $this->mapLocation( $location, $content, $section );
+
+            if ( $location->id == $content->versionInfo->contentInfo->mainLocationId )
+            {
+                $mainLocation = $location;
+            }
+
+            if ( !$location->hidden && !$location->invisible )
+            {
+                $isSomeLocationVisible = true;
+            }
+        }
+
+        // UserGroups and Users are Content, but permissions cascade is achieved through
+        // Locations hierarchy. We index all ancestor Location Content ids of all
+        // Locations of an owner.
+        $ancestorLocationsContentIds = $this->getAncestorLocationsContentIds(
+            $content->versionInfo->contentInfo->ownerId
+        );
+        // Add owner user id as it can also be considered as user group.
+        $ancestorLocationsContentIds[] = $content->versionInfo->contentInfo->ownerId;
+
+        $fields = array(
+            new Field(
+                'id',
+                'content' . $content->versionInfo->contentInfo->id,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'document_type',
+                'content',
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'type',
+                $content->versionInfo->contentInfo->contentTypeId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'version',
+                $content->versionInfo->versionNo,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'status',
+                $content->versionInfo->status,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'name',
+                $content->versionInfo->contentInfo->name,
+                new FieldType\StringField()
+            ),
+            new Field(
+                'creator',
+                $content->versionInfo->creatorId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'owner',
+                $content->versionInfo->contentInfo->ownerId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'owner_user_group',
+                $ancestorLocationsContentIds,
+                new FieldType\MultipleIdentifierField()
+            ),
+            new Field(
+                'section',
+                $content->versionInfo->contentInfo->sectionId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'section_identifier',
+                $section->identifier,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'section_name',
+                $section->name,
+                new FieldType\StringField()
+            ),
+            new Field(
+                'remote_id',
+                $content->versionInfo->contentInfo->remoteId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'modified',
+                $content->versionInfo->contentInfo->modificationDate,
+                new FieldType\DateField()
+            ),
+            new Field(
+                'published',
+                $content->versionInfo->contentInfo->publicationDate,
+                new FieldType\DateField()
+            ),
+            new Field(
+                'language_code',
+                array_keys( $content->versionInfo->names ),
+                new FieldType\MultipleStringField()
+            ),
+            new Field(
+                'main_language_code',
+                $content->versionInfo->contentInfo->mainLanguageCode,
+                new FieldType\StringField()
+            ),
+            new Field(
+                'always_available',
+                $content->versionInfo->contentInfo->alwaysAvailable,
+                new FieldType\BooleanField()
+            ),
+            new Field(
+                'some_location_visible',
+                $isSomeLocationVisible,
+                new FieldType\BooleanField()
+            ),
+        );
+
+        if ( $mainLocation !== null )
+        {
+            $fields[] = new Field(
+                'main_location',
+                $mainLocation->id,
+                new FieldType\IdentifierField()
+            );
+            $fields[] = new Field(
+                'main_location_parent',
+                $mainLocation->parentId,
+                new FieldType\IdentifierField()
+            );
+            $fields[] = new Field(
+                'main_location_remote_id',
+                $mainLocation->remoteId,
+                new FieldType\IdentifierField()
+            );
+            $fields[] = new Field(
+                'main_location_visible',
+                !$mainLocation->hidden && !$mainLocation->invisible,
+                new FieldType\BooleanField()
+            );
+            $fields[] = new Field(
+                'main_location_path',
+                $mainLocation->pathString,
+                new FieldType\IdentifierField()
+            );
+            $fields[] = new Field(
+                'main_location_depth',
+                $mainLocation->depth,
+                new FieldType\IntegerField()
+            );
+            $fields[] = new Field(
+                'main_location_priority',
+                $mainLocation->priority,
+                new FieldType\IntegerField()
+            );
+        }
+
+        $objectStateIds = array();
+        foreach ( $this->objectStateHandler->loadAllGroups() as $objectStateGroup )
+        {
+            $objectStateIds[] = $this->objectStateHandler->getContentState(
+                $content->versionInfo->contentInfo->id,
+                $objectStateGroup->id
+            )->id;
+        }
+
+        $contentType = $this->contentTypeHandler->load( $content->versionInfo->contentInfo->contentTypeId );
+        $fields[] = new Field(
+            'group',
+            $contentType->groupIds,
+            new FieldType\MultipleIdentifierField()
+        );
+
+        $fields[] = new Field(
+            'object_state',
+            $objectStateIds,
+            new FieldType\MultipleIdentifierField()
+        );
+
+        $fieldSets = $this->mapContentFields( $content, $contentType );
+
+        $documents = array();
+
+        foreach ( $fieldSets as $languageCode => $translationFields )
+        {
+            $translationFields[] = new Field(
+                'meta_indexed_language_code',
+                $languageCode,
+                new FieldType\StringField()
+            );
+            $translationFields[] = new Field(
+                'meta_indexed_is_main_translation',
+                ( $languageCode === $content->versionInfo->contentInfo->mainLanguageCode ),
+                new FieldType\BooleanField()
+            );
+            $translationFields[] = new Field(
+                'meta_indexed_is_main_translation_and_always_available',
+                (
+                    ( $languageCode === $content->versionInfo->contentInfo->mainLanguageCode ) &&
+                    $content->versionInfo->contentInfo->alwaysAvailable
+                ),
+                new FieldType\BooleanField()
+            );
+
+            $documents[] = new Document(
+                array(
+                    "languageCode" => $languageCode,
+                    "fields" => array_merge( $fields, $translationFields ),
+                    "documents" => $locationDocuments,
+                )
+            );
+        }
+
+        return $documents;
+    }
+
+    /**
+     * Maps given Content to a Document.
+     *
+     * @param \eZ\Publish\SPI\Persistence\Content $content
+     * @param \eZ\Publish\SPI\Persistence\Content\Type $contentType
+     *
+     * @return \eZ\Publish\SPI\Search\Field[][]
+     */
+    protected function mapContentFields( Content $content, ContentType $contentType )
+    {
+        $fieldSets = array();
+
+        foreach ( $content->fields as $field )
+        {
+            foreach ( $contentType->fieldDefinitions as $fieldDefinition )
+            {
+                if ( $fieldDefinition->id !== $field->fieldDefinitionId )
+                {
+                    continue;
+                }
+
+                $fieldType = $this->fieldRegistry->getType( $field->type );
+                foreach ( $fieldType->getIndexData( $field ) as $indexField )
+                {
+                    if ( $indexField->value === null )
+                    {
+                        continue;
+                    }
+
+                    $fieldSets[$field->languageCode][] = new Field(
+                        $this->fieldNameGenerator->getName(
+                            $indexField->name,
+                            $fieldDefinition->identifier,
+                            $contentType->identifier
+                        ),
+                        $indexField->value,
+                        $indexField->type
+                    );
+                }
+            }
+        }
+
+        return $fieldSets;
+    }
+
+    /**
+     * Maps given Location to a Document.
+     *
+     * Returned Document represents a "parent" Location document searchable with Location Search.
+     *
+     * @param \eZ\Publish\SPI\Persistence\Content\Location $location
+     * @param \eZ\Publish\SPI\Persistence\Content $content
+     * @param \eZ\Publish\SPI\Persistence\Content\Section $section
+     *
+     * @return \eZ\Publish\SPI\Search\Document
+     */
+    protected function mapLocation( Location $location, Content $content, Section $section )
+    {
+        $fields = array(
+            new Field(
+                'id',
+                'location' . $location->id,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'document_type',
+                'location',
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'priority',
+                $location->priority,
+                new FieldType\IntegerField()
+            ),
+            new Field(
+                'hidden',
+                $location->hidden,
+                new FieldType\BooleanField()
+            ),
+            new Field(
+                'invisible',
+                $location->invisible,
+                new FieldType\BooleanField()
+            ),
+            new Field(
+                'remote_id',
+                $location->remoteId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'content_id',
+                $location->contentId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'parent_id',
+                $location->parentId,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'path_string',
+                $location->pathString,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'depth',
+                $location->depth,
+                new FieldType\IntegerField()
+            ),
+            new Field(
+                'sort_field',
+                $location->sortField,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'sort_order',
+                $location->sortOrder,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'is_main_location',
+                ( $location->id == $content->versionInfo->contentInfo->mainLocationId ),
+                new FieldType\BooleanField()
+            ),
+            // Note: denormalized Content data is prefixed with 'content_' to avoid
+            // conflicts when using parent filter
+            new Field(
+                'content_name',
+                $content->versionInfo->contentInfo->name,
+                new FieldType\StringField()
+            ),
+            new Field(
+                'content_section_identifier',
+                $section->identifier,
+                new FieldType\IdentifierField()
+            ),
+            new Field(
+                'content_section_name',
+                $section->name,
+                new FieldType\StringField()
+            ),
+            new Field(
+                'content_modified',
+                $content->versionInfo->contentInfo->modificationDate,
+                new FieldType\DateField()
+            ),
+            new Field(
+                'content_published',
+                $content->versionInfo->contentInfo->publicationDate,
+                new FieldType\DateField()
+            ),
+        );
+
+        return new Document(
+            array(
+                "fields" => $fields
+            )
+        );
+    }
+
+    /**
+     * Returns Content ids of all ancestor Locations of all Locations
+     * of a Content with given $contentId.
+     *
+     * Used to determine user groups of a user with $contentId.
+     *
+     * @param int|string $contentId
+     *
+     * @return array
+     */
+    protected function getAncestorLocationsContentIds( $contentId )
+    {
+        $locations = $this->locationHandler->loadLocationsByContent( $contentId );
+        $ancestorLocationContentIds = array();
+        $ancestorLocationIds = array();
+
+        foreach ( $locations as $location )
+        {
+            $locationIds = explode( "/", trim( $location->pathString, "/" ) );
+            // Remove Location of Content with $contentId
+            array_pop( $locationIds );
+            // Remove Root Location id (id==1 in legacy DB)
+            array_shift( $locationIds );
+
+            $ancestorLocationIds = array_merge( $ancestorLocationIds, $locationIds );
+        }
+
+        foreach ( array_unique( $ancestorLocationIds ) as $locationId )
+        {
+            $location = $this->locationHandler->load( $locationId );
+
+            $ancestorLocationContentIds[$location->contentId] = true;
+        }
+
+        return array_keys( $ancestorLocationContentIds );
+    }
+}
